requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-aws>=0.2.0
langchain-community>=0.3.0
boto3>=1.35.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
numpy>=1.26.0
pydantic>=2.0.0